+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# <a href='https://github.com/mrjackwills/belugasnooze_vue/releases/tag/v0.3.10'>v0.3.10</a>
+### 2026-04-23
+
+### Chores
++ GitHub workflow updated, [f102bd9c](https://github.com/mrjackwills/belugasnooze_vue/commit/f102bd9c2486303ec97f279e83e146ce932cb090)
++ dependencies updated, [8d251ea3](https://github.com/mrjackwills/belugasnooze_vue/commit/8d251ea3712cb3525fc023717b84980863cc3803)
+
+### Fixes
++ TimeZone margin, [957ab6c5](https://github.com/mrjackwills/belugasnooze_vue/commit/957ab6c53fb808901bac74d7c812a30ab7c102c6)
+
 # <a href='https://github.com/mrjackwills/belugasnooze_vue/releases/tag/v0.3.9'>v0.3.9</a>
 ### 2026-03-07